Fix provisioning crash: RouterOS returns syslog port as string
LoggingAction.remote_port was typed as Option<u16> without the
ros_u32_opt deserializer. RouterOS returns "514" as a string,
causing deserialization failure on the Setup page.

Changed to Option<u32> with ros_u32_opt, matching the pattern
used by all other numeric fields in mikrotik-core resources.

+++ crates/ion-drift-web/src/provision.rs
@@ -219,7 +219,7 @@ pub async fn generate_plan(
     match existing_action {
         Some(action) => {
             let remote_matches = action.remote.as_deref() == Some(&config.syslog_host);
-            let port_matches = action.remote_port == Some(config.syslog_port);
+            let port_matches = action.remote_port == Some(config.syslog_port as u32);
             let src_matches = action.src_address.as_deref() == Some(&config.router_source_ip);
 
             if remote_matches && port_matches && src_matches {
